requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-cors-headers==3.7.0
uwsgi
gunicorn
uvicorn
pyyaml
Pillow==8.3.1
pandas==1.3.1
numpy==1.20.3